requirements/base.txt
django==4.0.5
django-environ==0.8.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.3.2
django-phonenumber-field== 6.1.0
phonenumbers==8.12.50
drf-yasg==1.20.0
Pillow==9.1.1
argon2-cffi==21.3.0
pytz==2022.1
